requirements.txt
maafw
requests
psutil
asyncify
qasync
cryptography
semver
pandas